Tidy up Game (#5)
Summary:
After 'finishing' the game yesterday, we realised there were certain exception-handling-things we had forgotten to include. Namely the logic that checked if a square was free before placing a symbol and checking if the user input was within the expected range of 0-8. We also decided to add prompts that allows the user to know the status of the game (i.e. if it's their turn). We added these extra requirements to the spec file too.

To-do:
We could tidy up some more by splitting up the TicTacToe class into smaller classes that abide to the Single Responsibility principle.